+++ tests/integration/scie/test_pex_scie.py
@@ -163,6 +163,10 @@ def test_multiple_platforms(tmpdir):
                 "--platform",
                 "linux-armv7l-cp-311-cp311",
                 "--platform",
+                "linux-ppc64le-cp-312-cp312",
+                "--platform",
+                "linux-s390x-cp-313-cp313",
+                "--platform",
                 "linux-x86_64-cp-310-cp310",
                 "--platform",
                 "macosx-10.9-arm64-cp-311-cp311",
@@ -175,6 +179,8 @@ def test_multiple_platforms(tmpdir):
     python_version_by_platform = {
         SciePlatform.LINUX_AARCH64: "3.9",
         SciePlatform.LINUX_ARMV7L: "3.11",
+        SciePlatform.LINUX_PPC64LE: "3.12",
+        SciePlatform.LINUX_S390X: "3.13",
         SciePlatform.LINUX_X86_64: "3.10",
         SciePlatform.MACOS_AARCH64: "3.11",
         SciePlatform.MACOS_X86_64: "3.12",
@@ -232,13 +238,15 @@ def test_multiple_platforms(tmpdir):
         expected_platforms=(
             SciePlatform.LINUX_AARCH64,
             SciePlatform.LINUX_ARMV7L,
+            SciePlatform.LINUX_PPC64LE,
+            SciePlatform.LINUX_S390X,
             SciePlatform.LINUX_X86_64,
             SciePlatform.MACOS_AARCH64,
             SciePlatform.MACOS_X86_64,
         ),
     )
 
-    # Now restrict the PEX's implied natural platform set of 5 down to 2 or 3 using
+    # Now restrict the PEX's implied natural platform set of 7 down to 2 or 3 using
     # `--scie-platform`.
     restricted_platforms_output_dir = os.path.join(str(tmpdir), "restricted-platforms")
     create_scies(
@@ -302,7 +310,7 @@ def test_specified_science_binary(tmpdir):
 
     local_science_binary = os.path.join(str(tmpdir), "science")
     with open(local_science_binary, "wb") as write_fp, URLFetcher().get_body_stream(
-        "https://github.com/a-scie/lift/releases/download/v0.9.0/{binary}".format(
+        "https://github.com/a-scie/lift/releases/download/v0.10.0/{binary}".format(
             binary=SciePlatform.CURRENT.qualified_binary_name("science")
         )
     ) as read_fp:
@@ -346,7 +354,7 @@ def test_specified_science_binary(tmpdir):
         cached_science_binaries
     ), "Expected the local science binary to be used but not cached."
     assert (
-        "0.9.0"
+        "0.10.0"
         == subprocess.check_output(args=[local_science_binary, "--version"]).decode("utf-8").strip()
     )
 
@@ -394,6 +402,8 @@ def test_custom_lazy_urls(tmpdir):
                         "x86_64-apple-darwin",
                         "aarch64-unknown-linux-gnu",
                         "armv7-unknown-linux-gnueabihf",
+                        "ppc64le-unknown-linux-gnu",
+                        "s390x-unknown-linux-gnu",
                         "x86_64-unknown-linux-gnu",
                     )
                 }
@@ -419,6 +429,10 @@ def test_custom_lazy_urls(tmpdir):
         expected_platform = "aarch64-unknown-linux-gnu"
     elif SciePlatform.CURRENT is SciePlatform.LINUX_ARMV7L:
         expected_platform = "armv7-unknown-linux-gnueabihf"
+    elif SciePlatform.CURRENT is SciePlatform.LINUX_PPC64LE:
+        expected_platform = "ppc64le-unknown-linux-gnu"
+    elif SciePlatform.CURRENT is SciePlatform.LINUX_S390X:
+        expected_platform = "s390x-unknown-linux-gnu"
     elif SciePlatform.CURRENT is SciePlatform.LINUX_X86_64:
         expected_platform = "x86_64-unknown-linux-gnu"
     elif SciePlatform.CURRENT is SciePlatform.MACOS_AARCH64: